third_party/WebKit/LayoutTests/fast/notifications/notifications-permission-expected.txt

Tests that the Notification.permission static property exists.

PASS window.Notification is defined.
PASS window.Notification.permission is defined.
PASS successfullyParsed is true

TEST COMPLETE
